Avoid calling Bundler.require in library code
This was causing us problems [1] in an application which uses Bundler
when we upgraded to Bundler >= v1.13.0.

I think the right solution is something along the lines of #3, but I'm
trying this as a temporary quick fix.

Note that this branch is based off the v1.0.0 tag which is the version
of the gem that we are currently using in our app.

[1]: alphagov/manuals-publisher#816